refactor code into components
