+++ lib/components/budget_year.dart
@@ -1,0 +1,37 @@
+import 'package:flutter/material.dart';
+
+class TitleButtonComponent extends StatelessWidget {
+  final String topTitle;
+  final String buttonText;
+
+  const TitleButtonComponent({
+    Key? key,
+    required this.topTitle,
+    required this.buttonText,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Align(
+      alignment: Alignment.center,
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+        children: [
+          Text(
+            topTitle,
+            style: const TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
+          ),
+          ElevatedButton(
+            onPressed: () {},
+            style: ElevatedButton.styleFrom(
+                elevation: 0,
+                backgroundColor: Theme.of(context).colorScheme.secondary,
+                shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(12))),
+            child: Text(buttonText),
+          ),
+        ],
+      ),
+    );
+  }
+}
