+++ lib/pages/home_page.dart
@@ -91,38 +91,4 @@ class _HomePageState extends State<HomePage> {
   }
 }
 
-class TitleButtonComponent extends StatelessWidget {
-  final String topTitle;
-  final String buttonText;
 
-  const TitleButtonComponent({
-    Key? key,
-    required this.topTitle,
-    required this.buttonText,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Align(
-      alignment: Alignment.center,
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          Text(
-            topTitle,
-            style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
-          ),
-          ElevatedButton(
-            onPressed: () {},
-            child: Text(buttonText),
-            style: ElevatedButton.styleFrom(
-                elevation: 0,
-                backgroundColor: Theme.of(context).colorScheme.secondary,
-                shape: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.circular(12))),
-          ),
-        ],
-      ),
-    );
-  }
-}
